+++ src/app/login/page.tsx
@@ -4,12 +4,12 @@ import { useForm, SubmitHandler } from 'react-hook-form';
 
 import Link from 'next/link';
 
+import useLogin from '@/apis/queryHooks/Auth/useLogin';
 import GoogleIcon from '@/assets/svg/google.svg';
 import NaverIcon from '@/assets/svg/naver.svg';
 import CommonButton from '@/components/CommonButton';
 import CommonInput from '@/components/CommonInput';
 import MaxLayout from '@/components/MaxLayout';
-import useLogin from '@/hooks/useLogin';
 import sha256 from '@/utils/sha256';
 
 import * as S from './Login.css';
@@ -26,7 +26,7 @@ function Home() {
     formState: { errors },
   } = useForm<Inputs>();
 
-  const login = useLogin();
+  const { mutate: login } = useLogin();
   const onSubmit: SubmitHandler<Inputs> = async data => {
     const newPassword = await sha256(data.passwordRequired);
